backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.28
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
anthropic>=0.20.0
litellm>=1.34.0
pypdf>=4.1.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
redis>=5.0.3
celery>=5.3.6
pytest>=8.1.1
pytest-asyncio>=0.23.5
httpx>=0.27.0
python-dotenv>=1.0.1
websockets>=12.0
tenacity>=8.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.1.0
greenlet>=3.0.0
